Added sounds to animals and main menu
I'll continue to improve the sounds. Also, right now the animal's
attacking is broken. If you are anywhere within the circle collision,
you take damage, so the first level is almost impossible.
I did make improvements to the health and lives so it reset after death,
but I couldn't find where to do it when you died to an animal