Added Options, menu, sound, and new function in PlayerController
Got all sign buttons working and added an options menu. It does nothing
yet. Added a grunt and knife swing sound effect for the player.
I also added a new function in PlayerController for when the player
takes damage. See if it will work for when the animals do damage